# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Phone Manager API
  version: "0.10.0"
  description: |
    REST API for the Phone Manager mobile application.
    Handles device registration, location tracking, group management,
    user authentication, and trip tracking for family/friends location sharing.

    ## Authentication

    The API supports multiple authentication methods:

    ### JWT Authentication (User APIs)
    User-facing endpoints use JWT Bearer tokens obtained via `/api/v1/auth/login`.
    Include the token in the `Authorization` header: `Bearer <access_token>`.

    ### API Key Authentication (Device APIs)
    Device and admin endpoints require an API key via the `X-API-Key` header.
    API keys use the format `pm_<key>`.

    ## Rate Limiting

    API requests are rate-limited:
    - Per API key: 100 requests/minute (default)
    - Forgot password: 5 requests/hour per IP
    - Request verification: 3 requests/hour per IP

    When rate limited, responses include a `Retry-After` header.

    ## Idempotency

    Location upload endpoints support idempotent requests via the `Idempotency-Key` header.
  contact:
    name: Phone Manager Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://loc.rlt.sk
    description: Production
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for device/admin authentication. Format: pm_<key>"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ValidationDetail"
    ValidationDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string

    # ==========================================
    # Organization API Key Schemas
    # ==========================================
...
    ApiKeyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key_prefix:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
...
    UpdateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 255

    # ==========================================
    # Organization Invitation Schemas
    # ==========================================
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /api/admin/v1/organizations/{org_id}/api-keys/{key_id}:
    get:
      tags: [Organization API Keys]
      summary: Get API key details
      operationId: getOrgApiKey
      security:
        - ApiKeyAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: key_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: API key details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Organization API Keys]
      summary: Update API key
      operationId: updateOrgApiKey
      security:
        - ApiKeyAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: key_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApiKeyRequest"
      responses:
        "200":
          description: API key updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Organization API Keys]
      summary: Revoke API key
      operationId: revokeOrgApiKey
      security:
        - ApiKeyAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: key_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: API key revoked
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================
  # Organization Member Invitations
  # ==========================================
...